move the C bridge into the py package
