+++ aggregator/singleton.go
@@ -1,0 +1,11 @@
+package aggregator
+
+var _aggregator Aggregator
+
+func Get() Aggregator {
+	return _aggregator
+}
+
+func Set(aggregatorInstance Aggregator) {
+	_aggregator = aggregatorInstance
+}
